// class-room-sauvegarde/11/class-room-frontend/src/app/api/model/niveauDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EtablissementProfesseurDto } from './etablissementProfesseurDto';
import { ClasseDto } from './classeDto';


export interface NiveauDto { 
    idNiveau?: number;
    libelle?: string;
    etablissementProfesseur?: EtablissementProfesseurDto;
    classes?: Set<ClasseDto>;
}