// class-room-sauvegarde/11/class-room-frontend/src/app/api/model/etablissementProfesseurDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProfesseurDto } from './professeurDto';
import { AnneeScolaireDto } from './anneeScolaireDto';
import { EtablissementDto } from './etablissementDto';


export interface EtablissementProfesseurDto { 
    idEtablissementProfesseur?: number;
    professeur?: ProfesseurDto;
    etablissement?: EtablissementDto;
    anneeScolaire?: AnneeScolaireDto;
}